requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
crispy-bootstrap4==2022.1
Django==3.2.23
django-appconf==1.0.6
django-bootstrap4==23.1
django-crispy-forms==2.0
django-environ==0.11.2
django-select2==7.11.1
django-tinymce4-lite==1.8.0
django_debug_toolbar==3.8.1
environ==1.0
faiss-cpu==1.7.4
filelock==3.12.2
fsspec==2023.1.0
git-filter-repo==2.45.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hangul-romanize==0.1.0
hpack==3.0.0
hstspreload==2024.6.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.16.4
hyperframe==5.2.0
idna==2.10
importlib-metadata==2.1.3
joblib==1.3.2
jsmin==3.0.1
ko-sentence-transformers==0.3
nltk==3.8.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
rfc3986==1.5.0
safetensors==0.4.3
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.4.1
sqlparse==0.4.4
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.4
transformers==4.30.2
typing_extensions==4.7.1
urllib3==2.0.7
zipp==3.15.0